test(query-core/notifyManager): move 'batchCalls' signature type test to 'test-d'
